requirements.dev.txt
--requirement=requirements.txt
django-debug-toolbar==1.5
django-silk==0.6.2